chore: refactor name
